Venue planning by kMeans clustering Jupyter Notebook
This Jupyter Notebook shows the work done by attaining, analyzing, and running kMeans clustering on Foursquare location of venues surrounding the Peabody Hotel in downtown Memphis, Tennessee. This project serves as the final capstone project for the IBM Professional Data Science course. As the final part of the course, I have been tasked with formulating a (business) problem that can be solved using machine learning techniques and Foursquare location data. The problem I have chosen to solve is vacation planning when visiting the city of Memphis. By using the Foursquare "explore" feature to get popular venues within a specified radius of a pre-determined location, combined with using the kMeans clustering algorithm, this project will determine the best spots to visit in Memphis for anyone staying in Memphis for 3, 5, or 7 days.